+++ proto/js/adminapi_pb.js
@@ -7432,7 +7432,7 @@ proto.adminapi.esm.UpdateUserStatusGemBalanceRequest.deserializeBinaryFromReader
       msg.setStatus(value);
       break;
     case 3:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setGemBalance(value);
       break;
     default:
@@ -7480,7 +7480,7 @@ proto.adminapi.esm.UpdateUserStatusGemBalanceRequest.serializeBinaryToWriter = f
   }
   f = message.getGemBalance();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       3,
       f
     );
@@ -7525,7 +7525,7 @@ proto.adminapi.esm.UpdateUserStatusGemBalanceRequest.prototype.setStatus = funct
 
 
 /**
- * optional int64 gem_balance = 3;
+ * optional int32 gem_balance = 3;
  * @return {number}
  */
 proto.adminapi.esm.UpdateUserStatusGemBalanceRequest.prototype.getGemBalance = function() {
@@ -9237,7 +9237,7 @@ proto.adminapi.esm.UserDetail.deserializeBinaryFromReader = function(msg, reader
       msg.setStatus(value);
       break;
     case 11:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setGemBalance(value);
       break;
     case 12:
@@ -9429,7 +9429,7 @@ proto.adminapi.esm.UserDetail.serializeBinaryToWriter = function(message, writer
   }
   f = message.getGemBalance();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       11,
       f
     );
@@ -9772,7 +9772,7 @@ proto.adminapi.esm.UserDetail.prototype.setStatus = function(value) {
 
 
 /**
- * optional int64 gem_balance = 11;
+ * optional int32 gem_balance = 11;
  * @return {number}
  */
 proto.adminapi.esm.UserDetail.prototype.getGemBalance = function() {
@@ -31084,7 +31084,7 @@ proto.adminapi.esm.AddSubscriptionRequest.deserializeBinaryFromReader = function
       msg.setQuantity(value);
       break;
     case 8:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setOneTimeGem(value);
       break;
     case 9:
@@ -31096,7 +31096,7 @@ proto.adminapi.esm.AddSubscriptionRequest.deserializeBinaryFromReader = function
       msg.setOneTimeIsFirstonly(value);
       break;
     case 11:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setDailyGem(value);
       break;
     case 12:
@@ -31187,7 +31187,7 @@ proto.adminapi.esm.AddSubscriptionRequest.serializeBinaryToWriter = function(mes
   }
   f = message.getOneTimeGem();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       8,
       f
     );
@@ -31208,7 +31208,7 @@ proto.adminapi.esm.AddSubscriptionRequest.serializeBinaryToWriter = function(mes
   }
   f = message.getDailyGem();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       11,
       f
     );
@@ -31357,7 +31357,7 @@ proto.adminapi.esm.AddSubscriptionRequest.prototype.setQuantity = function(value
 
 
 /**
- * optional int64 one_time_gem = 8;
+ * optional int32 one_time_gem = 8;
  * @return {number}
  */
 proto.adminapi.esm.AddSubscriptionRequest.prototype.getOneTimeGem = function() {
@@ -31411,7 +31411,7 @@ proto.adminapi.esm.AddSubscriptionRequest.prototype.setOneTimeIsFirstonly = func
 
 
 /**
- * optional int64 daily_gem = 11;
+ * optional int32 daily_gem = 11;
  * @return {number}
  */
 proto.adminapi.esm.AddSubscriptionRequest.prototype.getDailyGem = function() {
@@ -31709,7 +31709,7 @@ proto.adminapi.esm.UpdateSubscriptionRequest.deserializeBinaryFromReader = funct
       msg.setQuantity(value);
       break;
     case 9:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setOneTimeGem(value);
       break;
     case 10:
@@ -31721,7 +31721,7 @@ proto.adminapi.esm.UpdateSubscriptionRequest.deserializeBinaryFromReader = funct
       msg.setOneTimeIsFirstonly(value);
       break;
     case 12:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setDailyGem(value);
       break;
     case 13:
@@ -31819,7 +31819,7 @@ proto.adminapi.esm.UpdateSubscriptionRequest.serializeBinaryToWriter = function(
   }
   f = message.getOneTimeGem();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       9,
       f
     );
@@ -31840,7 +31840,7 @@ proto.adminapi.esm.UpdateSubscriptionRequest.serializeBinaryToWriter = function(
   }
   f = message.getDailyGem();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       12,
       f
     );
@@ -32007,7 +32007,7 @@ proto.adminapi.esm.UpdateSubscriptionRequest.prototype.setQuantity = function(va
 
 
 /**
- * optional int64 one_time_gem = 9;
+ * optional int32 one_time_gem = 9;
  * @return {number}
  */
 proto.adminapi.esm.UpdateSubscriptionRequest.prototype.getOneTimeGem = function() {
@@ -32061,7 +32061,7 @@ proto.adminapi.esm.UpdateSubscriptionRequest.prototype.setOneTimeIsFirstonly = f
 
 
 /**
- * optional int64 daily_gem = 12;
+ * optional int32 daily_gem = 12;
  * @return {number}
  */
 proto.adminapi.esm.UpdateSubscriptionRequest.prototype.getDailyGem = function() {
@@ -32999,7 +32999,7 @@ proto.adminapi.esm.SubscriptionDetail.deserializeBinaryFromReader = function(msg
       msg.setQuantity(value);
       break;
     case 9:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setOneTimeGem(value);
       break;
     case 10:
@@ -33011,7 +33011,7 @@ proto.adminapi.esm.SubscriptionDetail.deserializeBinaryFromReader = function(msg
       msg.setOneTimeIsFirstonly(value);
       break;
     case 12:
-      var value = /** @type {number} */ (reader.readInt64());
+      var value = /** @type {number} */ (reader.readInt32());
       msg.setDailyGem(value);
       break;
     case 13:
@@ -33109,7 +33109,7 @@ proto.adminapi.esm.SubscriptionDetail.serializeBinaryToWriter = function(message
   }
   f = message.getOneTimeGem();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       9,
       f
     );
@@ -33130,7 +33130,7 @@ proto.adminapi.esm.SubscriptionDetail.serializeBinaryToWriter = function(message
   }
   f = message.getDailyGem();
   if (f !== 0) {
-    writer.writeInt64(
+    writer.writeInt32(
       12,
       f
     );
@@ -33297,7 +33297,7 @@ proto.adminapi.esm.SubscriptionDetail.prototype.setQuantity = function(value) {
 
 
 /**
- * optional int64 one_time_gem = 9;
+ * optional int32 one_time_gem = 9;
  * @return {number}
  */
 proto.adminapi.esm.SubscriptionDetail.prototype.getOneTimeGem = function() {
@@ -33351,7 +33351,7 @@ proto.adminapi.esm.SubscriptionDetail.prototype.setOneTimeIsFirstonly = function
 
 
 /**
- * optional int64 daily_gem = 12;
+ * optional int32 daily_gem = 12;
  * @return {number}
  */
 proto.adminapi.esm.SubscriptionDetail.prototype.getDailyGem = function() {
